[refactor] move code into jsonUtils file
