+++ src/utils/ressources/jsonUtils.ts
@@ -1,3 +1,5 @@
+import {roomDetailsType, submissionForm} from "./types";
+
 export function findAllKeysForSubmission(obj: object) {
 	let results: any[] = [];
 
@@ -32,4 +34,30 @@ export function compareVersions(oldVersion: object[], newVersion: object[], actu
 	} else {
 		return version[0] + "." + version[1] + "." + (+version[2] + 1);
 	}
+}
+
+export function readOrEditHazard(room: roomDetailsType, action: string, currentForm: any, withForm: boolean): submissionForm[] {
+	const subForm: submissionForm[] = [];
+	room.hazards.forEach(h => {
+		try {
+			const category = h.hazard_form_history.hazard_form.hazard_category.hazard_category_name;
+			//if (category == selectedHazardCategory) {
+			const childrenList: submissionForm[] = [];
+			h.children.forEach(child => {
+				childrenList.push({id: child.id, submission: JSON.parse(child.submission),
+					form: withForm ? (action == 'Read' ? JSON.parse(child.hazard_form_child_history.form) : JSON.parse(child.hazard_form_child_history.hazard_form_child.form)) : {}});
+			})
+			subForm.push({id: h.id, submission: JSON.parse(h.submission), form: withForm ? (action == 'Read' ? JSON.parse(h.hazard_form_history.form) : currentForm) : {},
+				children: childrenList, room: room, category: category});
+			//}
+		} catch (error) {
+			console.error(error);
+		}
+	});
+	return subForm;
+}
+
+export function splitCamelCase(str: string) {
+	const label = str.replace(/([a-z])([A-Z])/g, '$1 $2') // Insert a space between lowercase and uppercase letters
+	return label.charAt(0).toUpperCase() + label.slice(1);
 }
